pyproject.toml
[tool.poetry]
authors = ["Bryan Matteson <bryan@matteson.dev>"]
description = ""
name = "cbconf"
packages = [{include = "cbconf", from = "src"}]
readme = "README.md"
version = "0.0.1"

[tool.poetry.dependencies]
pydantic = "^1.9.1"
python = "^3.8"
python-dotenv = {version = "^0.20.0", optional = true}

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
pytest = "^7.1.2"

[tool.poetry.extras]
dotenv = ["python-dotenv"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
use_parentheses = true

[tool.pycln]
all = true

[tool.black]
include = "\\.(py)$"
line-length = 120
target-version = ["py38"]

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
show_traceback = true

[tool.pyright]
exclude = ["**/__pycache__"]
ignore = []
include = ["src/cbconf"]
typeCheckingMode = "basic"

reportMissingImports = true
reportMissingTypeStubs = false

pythonPlatform = "All"
pythonVersion = "3.9"
